+++ src/components/LoginPage/index.js
@@ -1,0 +1,3 @@
+import LoginPage from "./Login";
+
+export default LoginPage;
