+++ src/components/RegisterPage/index.js
@@ -1,0 +1,3 @@
+import RegisterPage from "./Register";
+
+export default RegisterPage;
